@@ -1,18 +1,36 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
+use Illuminate\Database\Eloquent\Factories\Factory;
 use App\Models\Post;
-use Faker\Generator as Faker;
+use App\User;
 
-$factory->define(Post::class, function (Faker $faker) {
-    return [
-        'content' => $faker->text($maxNbChars = 255),
-    ];
-});
+class PostFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Post::class;
 
-$factory->state(Post::class, 'withAuthor', function (Faker $faker) {
-    return [
-        'author_id' => factory(App\User::class),
-    ];
-});
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'content' => $this->faker->text($maxNbChars = 255),
+        ];
+    }
+
+    public function with_author()
+    {
+        return $this->state([
+            'author_id' => User::factory(),
+        ]);
+    }
+}
