@@ -1,13 +1,29 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
+use Illuminate\Database\Eloquent\Factories\Factory;
 use App\Models\Technology;
-use Faker\Generator as Faker;
 
-$factory->define(Technology::class, function (Faker $faker) {
-    return [
-        'name' => $faker->unique()->word,
-        'label' => $faker->word
-    ];
-});
+class TechnologyFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Technology::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->unique()->word,
+            'label' => $this->faker->word
+        ];
+    }
+}
